no more triangle methods in views
